Main.py has a menu to use. You can run a single instance look at runSingle() to see what scrapers are used.
runInBack() lets you input the amount of hours you want the process to run, and the frequency that it runs. (Ex: 10 hours, 30 min intervals = instances)

look() allows you to get the count data for any stock in the root folder

changeSettings() is not operable rn

To avoid switch statement i made processCMD() i made a dict of commands that can be accessed with keys that represent the menu choices

printMenu() is the main printing function. Shows all the options and it is where processCMD is used